pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ha-mcp"
version = "0.1.0"
description = "MCP server for Home Assistant configuration management"
requires-python = ">= 3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "aiohttp>=3.9.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
ha-mcp = "ha_mcp.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ha_mcp"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov",
]
